+++ R/utils.R
@@ -39,6 +39,24 @@ bucket_location_to_flag <- function(bucket_location) {
   }
 }
 
+## Given an s3cmd path and a bucket location, will construct a flag
+## argument for s3cmd.  If it looks like the s3cmd is actually
+## pointing to an s4cmd, return empty string as s4cmd doesn't
+## support bucket location.
+bucket_location_to_flag <- function(bucket_location) {
+  if (using_s4cmd()) {
+    if (!identical(bucket_location, "US")) {
+        warning(paste0("Ignoring non-default bucket location ('",
+                       bucket_location,
+                       "') in s3mpi::s3.get since s4cmd was detected",
+                       "-- this might be a little slower but is safe to ignore."));
+    }
+    ""
+  } else {
+    paste("--bucket-location", bucket_location)
+  }
+}
+
 ## We use the [memoise](https://github.com/hadley/memoise) package to
 ## ensure this check only gets run once in a given R session. This
 ## means a user will have to restart R if they install s3cmd