+++ NEWS.md
@@ -1,3 +1,11 @@
+# Version 0.2.33
+
+* Other fixes for s4cmd.
+
+# Version 0.2.32
+* allow choice of storage format in `s3read` and `s3store`. Defaults to `RDS`,
+  and now you can chooose `CSV` or `table` for data frames.
+
 # Version 0.2.31
 
 * Make `s3store` work with s4cmd.